% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_folder.R
\name{create_project_folder}
\alias{create_project_folder}
\title{Set up template data sience project folder}
\usage{
create_project_folder(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{Path to directory in which to set up the template folder.}

\item{verbose}{Boolean, should a message be displayed after each step?
Default value is \code{TRUE}.}
}
\description{
Sets-up a template, ready-to-use folder for a new data science project.
}
\details{
Sets up the directory given in \code{path} as follows:
\itemize{
\item Creates folders \code{data/}, \code{output/} and \code{reports};
\item Creates or ammend the \code{.gitignore} file to exclude \code{targets} files
and folders;
\item Creates template RMarkdown files \code{make_targets_pipeline.Rmd} and
\code{data_cleaning.Rmd};
\item Creates template RMardown file \code{results.Rmd} in the \code{reports/} folder.
}
See \code{vignette("create_project_folder")} for more information on the
template and how to use it.
}
